Changed data sources, made map icons proportional to hotel rooms, added buttons and sliders
Added Hotel.geojson and adjusted script for hotel data.  Removed grocery store data and added crime data.

Changed the map popups from showing all of a record's (hotel) attributes to only showing its name and number of rooms.  Changed the map icons to be proportional to the number of rooms in each hotel.

Added buttons and a slider for the map to main.js and styles for them in styles.css.